requirements.txt
asgiref==3.8.1
Django==4.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
setuptools==75.6.0
sqlparse==0.5.2
uritemplate==4.1.1
